feat: add get TB admin token middleware
