+++ src/controller/user.controller.ts
@@ -4,9 +4,9 @@ import { Request, Response, NextFunction } from 'express';
 import checkStatusError from '../helpers/checkStatusError';
 import { autoLoginToGetTenantToken } from '../service/user';
 
-export async function autoLoginTenantAccount(_req: Request, res: Response, next: NextFunction) {
+export async function autoLoginTenantAccount(req: Request, res: Response, next: NextFunction) {
     try {
-        const response = await autoLoginToGetTenantToken();
+        const response = await autoLoginToGetTenantToken(req.headers.TBAdminToken as string);
         if (checkStatusError(response)) {
             return next({
                 status: response.status,
